Fix compilation

+++ SlipeServer.Server.TestTools/TestingServer.cs
@@ -73,8 +73,8 @@ public class TestingServer<TPlayer> : MtaServer<TPlayer>
 
     private void SetupSendPacketMocks()
     {
-        this.NetWrapperMock.Setup(x => x.SendPacket(It.IsAny<uint>(), It.IsAny<ushort>(), It.IsAny<Packet>()))
-            .Callback((uint address, ushort version, Packet packet) =>
+        this.NetWrapperMock.Setup(x => x.SendPacket(It.IsAny<ulong>(), It.IsAny<ushort>(), It.IsAny<Packet>()))
+            .Callback((ulong address, ushort version, Packet packet) =>
             {
                 this.sendPacketCalls.Add(new SendPacketCall()
                 {
@@ -88,13 +88,13 @@ public class TestingServer<TPlayer> : MtaServer<TPlayer>
             });
 
         this.NetWrapperMock.Setup(x => x.SendPacket(
-            It.IsAny<uint>(),
+            It.IsAny<ulong>(),
             It.IsAny<PacketId>(), 
             It.IsAny<ushort>(), 
             It.IsAny<byte[]>(), 
             It.IsAny<PacketPriority>(), 
             It.IsAny<PacketReliability>()
-        )).Callback((uint address, PacketId packetId, ushort version, byte[] data, PacketPriority priority, PacketReliability reliability) =>
+        )).Callback((ulong address, PacketId packetId, ushort version, byte[] data, PacketPriority priority, PacketReliability reliability) =>
             {
                 this.sendPacketCalls.Add(new SendPacketCall()
                 {
